+++ Areas/HumanResource/Controllers/PositionsController.cs
@@ -7,6 +7,9 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Standus_5_0.Areas.HumanResource.Models;
 using Standus_5_0.Data;
+using Standus_5_0.Enums;
+using Standus_5_0.Migrations.Allowance;
+using Standus_5_0.Services;
 
 namespace Standus_5_0.Areas.HumanResource.Controllers
 {
@@ -58,19 +61,38 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("PositionID,PositionName,Description,DepartmentID")] Position position)
+        public async Task<IActionResult> Create(Position position)
         {
+            ViewData["DepartmentID"] = new SelectList(_context.Department, "DepartmentID", "DepartmentName", position.DepartmentID);
+
             if (ModelState.IsValid)
             {
+                var positionType = _context.Position.Where(p => p.PositionName == position.PositionName)
+                    .Select(n => n.PositionName).FirstOrDefault();
+
+                if (positionType != null) {
+                    ModelState.AddModelError("PositionName", "This Position Exists !");
+                    return View(position);
+                }
+
                 _context.Add(position);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (position.PositionID  > 0)
+                {
+                    TempData["Alert"] = CommonServices.ShowAlert(Alerts.Success, "Data Saved.");
+                }
+                else
+                {
+                    TempData["Alert"] = CommonServices.ShowAlert(Alerts.Danger, "Unknown error.");
+                }
+                return RedirectToAction(nameof(Create));
             }
-            ViewData["DepartmentID"] = new SelectList(_context.Department, "DepartmentID", "DepartmentName", position.DepartmentID);
+            
             return View(position);
         }
 
         // GET: HumanResource/Positions/Edit/5
+        [HttpGet]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -78,6 +100,7 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
                 return NotFound();
             }
 
+           
             var position = await _context.Position.FindAsync(id);
             if (position == null)
             {
@@ -92,8 +115,10 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("PositionID,PositionName,Description,DepartmentID")] Position position)
+        public async Task<IActionResult> Edit(int id, Position position)
         {
+            ViewData["DepartmentID"] = new SelectList(_context.Department, "DepartmentID", "DepartmentName", position.DepartmentID);
+
             if (id != position.PositionID)
             {
                 return NotFound();
@@ -103,8 +128,28 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
             {
                 try
                 {
+                    
+                   // var positionType = _context.Position.Where(p => p.PositionName == position.PositionName)
+                   //.Select(n => n.PositionName).FirstOrDefault();
+
+                   // if (positionType != null)
+                   // {
+                   //     ModelState.AddModelError("PositionName", "This Position Exists !");
+                   //     return View(position);
+                   // }
+
                     _context.Update(position);
                     await _context.SaveChangesAsync();
+
+                    if (position.PositionID > 0)
+                    {
+                        TempData["Alert"] = CommonServices.ShowAlert(Alerts.Success, "Data Saved.");
+                    }
+                    else
+                    {
+                        TempData["Alert"] = CommonServices.ShowAlert(Alerts.Danger, "Unknown error.");
+                    }
+                    return RedirectToAction(nameof(Edit));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -119,7 +164,7 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["DepartmentID"] = new SelectList(_context.Department, "DepartmentID", "DepartmentName", position.DepartmentID);
+            
             return View(position);
         }
 
@@ -148,6 +193,15 @@ namespace Standus_5_0.Areas.HumanResource.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var position = await _context.Position.FindAsync(id);
+
+            var employ = _context.Employee.OfType<EmployementDetails>().FirstOrDefault(f => f.PositionID == id);
+
+            if (employ != null)
+            {
+                TempData["Alert"] = CommonServices.ShowAlert(Alerts.Danger, position.PositionName + " is in use. Select Inactivate !");
+                return RedirectToAction(nameof(Delete));
+            }
+
             if (position != null)
             {
                 _context.Position.Remove(position);
